✅ test : test validateDuplicateUser

+++ back-end/src/main/java/com/ohohchoo/domain/user/repository/UserRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface UserRepository extends JpaRepository<User, Integer> {
-    public Optional<User> findById(Integer id);
+public interface UserRepository extends JpaRepository<User, Long> {
+    public Optional<User> findById(Long id);
     public Optional<User> findByEmail(String email);
 }
